Refactor and simplify knight API
