+++ api/src/api/knight/fetch-knight.ts
@@ -1,0 +1,13 @@
+import { getKnightNTurn } from '../../services/knight/get-n-turn'
+import { positionInvalid } from '../../services/knight/position-helper'
+
+export const fetchKnight = (req, res) => {
+  const { currentPosition, turnNumber } = req.query
+
+  if (positionInvalid(currentPosition) || turnNumber === "0") {
+    return res.status(400).send({ 'error': 'Invalid position.' })
+  }
+  const possiblePositionList = getKnightNTurn(currentPosition, turnNumber)
+
+  return res.status(200).send({ 'data': possiblePositionList })
+}
